fix(frontend): correct constant-folding register targeting and de-duplicate purity helpers
- emitConst: write folded LOAD_CONST into the target register instead of
  leaving dst=-1, so folded results actually reach their destination
- foldBinary: require both operands known before identity folding, and
  restrict the left-operand identity to commutative ADD/MUL (prevents an
  uninitialized variant from being mistaken for the identity element,
  which broke recursive fib)
- hoist isPureBinary/isPureUnary into OpTraits.h as inline helpers to
  resolve a unity-build redefinition between CSEPass and LICMPass
- drop unused oldIdx parameters and a dead owner local to silence warnings

+++ src/LOICollectionA/frontend/ir/vm/VMExecCall.cpp
@@ -315,7 +315,7 @@ namespace LOICollection::frontend::ir {
 
                 const size_t paramBase = func->captures.size();
                 if (paramBase + static_cast<size_t>(func->argCount) >
-                    func->owner->methodBodies[func->bodyIndex]->slotCount) {
+                    static_cast<size_t>(func->owner->methodBodies[func->bodyIndex]->slotCount)) {
                     this->diagnostics.addError(this->currentLoc, "Lambda frame is too small for its parameters");
                     break;
                 }
